requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deep-translator==1.11.4
edge-tts==7.0.2
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
frozenlist==1.6.0
googletrans==4.0.0rc1
greenlet==3.2.2
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
moviepy==2.1.2
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==10.4.0
proglog==0.1.12
propcache==0.3.1
pydub==0.25.1
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytrends==4.9.2
pyttsx3==2.98
pytz==2025.2
pywin32==310
requests==2.32.3
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
srt==3.5.3
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
